project

+++ frontend/src/components/WorkoutDetails.js
@@ -1,13 +1,22 @@
 import { useWorkoutContext } from "../hooks/useWorkoutContext";
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = ({ workout }) => {
     const { dispatch } = useWorkoutContext();
-
+    const { isLoggedIn } = useAuthContext();
+    console.log(isLoggedIn);
+    
     const handleClick = async () => {
+        if (!isLoggedIn) {
+            return;
+        }
         const response = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${isLoggedIn.token}`
+            }
         })
         const json = await response.json();
